pyproject.toml
[project]
name = "sherlock-discord-bot"
version = "0.1.0"
description = "Discord chatbot para concurseiros brasileiros, powered by OpenRouter"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "asyncpg>=0.31.0",
    "dacite>=1.8.0",
    "discord.py>=2.6.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pypdf>=3.17.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=5.1.2",
    "tenacity>=9.1.2",
    "python-docx>=1.2.0",
    "beautifulsoup4>=4.14.3",
]

[project.scripts]
sherlock-bot = "src.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ============================================================================
# Ruff - Linting & Formatting
# ============================================================================
[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "ARG001", # unused function argument (some intentionally unused for API compat)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============================================================================
# Mypy - Type Checking
# ============================================================================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

# ============================================================================
# Pytest - Testing
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = "src"
addopts = [
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html",
    "--cov-report=xml:coverage.xml",
]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]
